+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### Added
 - Add `CellFilter::AnyOf(filters)` and `CellFilter::NoneOf(filters)` variants.
+- Implemented `ToRgbComponents` trait for `Color` to standardize extraction of RGB components.
 
 ### Fixed
 - `fx::translate()`: replace `todo!()` in cell_selection().
+- 16 and 256 color spaces no longer output black when interpolating to a different color.
 
 ## tachyonfx 0.1.0 - 2024-06-20